cleanup, moved some files to test
