+++ src/vslam/src/odometry/AlignmentRgbd.cpp
@@ -1,59 +1,48 @@
 
-
 #include <execution>
-#include <functional>
 #include <numeric>
 
-#include "AlignmentRgb.h"
+#include "AlignmentRgbd.h"
 #include "core/Feature2D.h"
 #include "core/random.h"
-#include "direct/interpolate.h"
-#include "direct/jacobians.h"
 #include "features/FeatureSelection.h"
+
 #include "utils/log.h"
+
+#include "interpolate.h"
+#include "jacobians.h"
 #define PERFORMANCE_RGBD_ALIGNMENT false
 #define LOG_NAME "direct_odometry"
 #define MLOG(level) CLOG(level, LOG_NAME)
-
+#define exec_policy std::execution::par_unseq
 namespace vslam {
-std::map<std::string, double> AlignmentRgb::defaultParameters() {
+std::map<std::string, double> AlignmentRgbd::defaultParameters() {
   return {{"nLevels", 4.0}, {"maxIterations", 100}, {"minParameterUpdate", 1e-4}, {"maxErrorIncrease", 1.5}};
 }
 
-AlignmentRgb::AlignmentRgb(const std::map<std::string, double> params) :
-    AlignmentRgb(params.at("nLevels"), params.at("maxIterations"), params.at("minParameterUpdate"), params.at("maxErrorIncrease")) {}
+AlignmentRgbd::AlignmentRgbd(const std::map<std::string, double> params) :
+    AlignmentRgbd(params.at("nLevels"), params.at("maxIterations"), params.at("minParameterUpdate"), params.at("maxErrorIncrease")) {}
 
-AlignmentRgb::AlignmentRgb(int nLevels, double maxIterations, double minParameterUpdate, double maxErrorIncrease) :
-    _weightFunction(std::make_shared<TDistribution<Constraint>>(5.0, 1e-3, 10)),
-    _nLevels(nLevels),
+AlignmentRgbd::AlignmentRgbd(int nLevels, double maxIterations, double minParameterUpdate, double maxErrorIncrease) :
+    _weightFunction(std::make_shared<TDistributionBivariate<Constraint>>(5.0, 1e-3, 10)),
+    _maxIterations(maxIterations),
+    _minParameterUpdate(minParameterUpdate),
+    _maxErrorIncrease(maxErrorIncrease) {
+  log::create(LOG_NAME);
+  for (int i = nLevels - 1; i >= 0; i--) {
+    _levels.push_back(i);
+  }
+}
+AlignmentRgbd::AlignmentRgbd(const std::vector<int> &levels, double maxIterations, double minParameterUpdate, double maxErrorIncrease) :
+    _weightFunction(std::make_shared<TDistributionBivariate<Constraint>>(5.0, 1e-3, 10)),
+    _levels(levels),
     _maxIterations(maxIterations),
     _minParameterUpdate(minParameterUpdate),
     _maxErrorIncrease(maxErrorIncrease) {
   log::create(LOG_NAME);
 }
 
-Pose AlignmentRgb::align(
-  Camera::ConstShPtr cam,
-  const cv::Mat &intensity0,
-  const cv::Mat &depth0,
-  const cv::Mat &intensity1,
-  const SE3d &guess,
-  const Mat6d &guessCovariance) {
-  auto f0 = std::make_shared<Frame>(intensity0, depth0, cam);
-  f0->computePyramid(_nLevels);
-  f0->computeDerivatives();
-  f0->computePcl();
-  auto featureSelection = std::make_shared<FeatureSelection<FiniteGradient>>(FiniteGradient{5, 0.01, 0.3, 0, 8.0}, 10, 4);
-  featureSelection->select(f0);
-  auto f1 = std::make_shared<Frame>(intensity1, cam);
-  f1->computePyramid(_nLevels);
-  f1->pose() = Pose(guess, guessCovariance);
-  Pose pose = align(f0, f1)->pose;
-  f0->removeFeatures();
-  return pose;
-}
-
-Pose AlignmentRgb::align(
+Pose AlignmentRgbd::align(
   Camera::ConstShPtr cam,
   const cv::Mat &intensity0,
   const cv::Mat &depth0,
@@ -62,38 +51,46 @@ Pose AlignmentRgb::align(
   const SE3d &guess,
   const Mat6d &guessCovariance) {
   auto f0 = std::make_shared<Frame>(intensity0, depth0, cam);
-  f0->computePyramid(_nLevels);
+  f0->computePyramid(*std::max_element(_levels.begin(), _levels.end()) + 1);
   f0->computeDerivatives();
   f0->computePcl();
   auto featureSelection = std::make_shared<FeatureSelection<FiniteGradient>>(FiniteGradient{5, 0.01, 0.3, 0, 8.0}, 10, 4);
 
   featureSelection->select(f0);
   auto f1 = std::make_shared<Frame>(intensity1, depth1, cam);
-  f1->computePyramid(_nLevels);
+  f1->computePyramid(*std::max_element(_levels.begin(), _levels.end()) + 1);
   f1->pose() = Pose(guess, guessCovariance);
   Pose pose = align(f0, f1)->pose;
   f0->removeFeatures();
   return pose;
 }
-AlignmentRgb::Results::UnPtr AlignmentRgb::align(Frame::ConstShPtr frame0, Frame::ConstShPtr frame1) {
+AlignmentRgbd::Results::UnPtr AlignmentRgbd::align(Frame::ConstShPtr frame0, Frame::ConstShPtr frame1) {
   return align(frame0->features(), frame1);
 }
-AlignmentRgb::Results::UnPtr AlignmentRgb::align(const Feature2D::VecConstShPtr &features, Frame::ConstShPtr frame1) {
+AlignmentRgbd::Results::UnPtr AlignmentRgbd::align(Frame::ConstShPtr frame0, Frame::ConstShPtr frame1, const Pose &prior) {
+  return align(frame0->features(), frame1, prior);
+}
+AlignmentRgbd::Results::UnPtr AlignmentRgbd::align(Feature2D::VecConstShPtr features, Frame::ConstShPtr frame1) {
+  // TODO require prior explicitly
+  return align(features, frame1, frame1->pose());
+}
+AlignmentRgbd::Results::UnPtr AlignmentRgbd::align(Feature2D::VecConstShPtr features, Frame::ConstShPtr frame1, const Pose &prior) {
+
   TIMED_SCOPE(timer, "align");
 
-  const Pose &prior = frame1->pose();
   SE3f pose = prior.SE3().cast<float>();
   Mat6f covariance;
   Frame::VecConstShPtr framesRef;
   std::transform(features.begin(), features.end(), std::back_inserter(framesRef), [](auto ft) { return ft->frame(); });
-
   std::vector<SE3f> motions(framesRef.size());
-  std::vector<Constraint::VecShPtr> constraints(_nLevels);
-  std::vector<NormalEquations> normalEquations(_nLevels);
-  std::vector<int> iterations(_nLevels);
-  std::vector<double> scale(_nLevels);
-  for (_level = _nLevels - 1; _level >= 0; _level--) {
+  std::vector<Constraint::VecShPtr> constraints(nLevels());
+  std::vector<NormalEquations> normalEquations(nLevels());
+  std::vector<int> iterations(nLevels());
+  std::vector<Mat2d> scale(nLevels());
+  for (const auto &l : _levels) {
+    _level = l;
     TIMED_SCOPE_IF(timerLevel, format("computeLevel{}", _level), PERFORMANCE_RGBD_ALIGNMENT);
+
     auto constraintsAll = setupConstraints(framesRef, features);
 
     std::string reason = "Max iterations exceeded";
@@ -109,15 +106,14 @@ AlignmentRgb::Results::UnPtr AlignmentRgb::align(const Feature2D::VecConstShPtr 
       constraints[_level] = computeResidualsAndJacobian(constraintsAll, frame1, motions);
 
       if (constraints[_level].size() < 6) {
-        reason = format("Not enough constraints: {}", constraints[_level].size());
+        reason = format("Not enough constraints: {} at {} iteration {}", constraints[_level].size(), _level, _iteration);
         MLOG(WARNING) << reason;
-        pose = SE3f();
         break;
       }
       {
         TIMED_SCOPE_IF(timer3, format("computeWeights{}", _level), PERFORMANCE_RGBD_ALIGNMENT);
         _weightFunction->computeWeights(constraints[_level]);
-        scale[_level] = _weightFunction->scale();
+        scale[_level] = _weightFunction->scale().cast<double>();
       }
 
       normalEquations[_level] = computeNormalEquations(constraints[_level]);
@@ -135,6 +131,11 @@ AlignmentRgb::Results::UnPtr AlignmentRgb::align(const Feature2D::VecConstShPtr 
 
       dx = normalEquations[_level].A.ldlt().solve(normalEquations[_level].b);
 
+      if (!dx.allFinite()) {
+        reason = format("NaN during optimization");
+        break;
+      }
+
       pose = SE3f::exp(-dx) * pose;
       covariance = normalEquations[_level].A.inverse();
 
@@ -142,16 +143,29 @@ AlignmentRgb::Results::UnPtr AlignmentRgb::align(const Feature2D::VecConstShPtr 
         reason = format("Minimum step size reached: {:.5f}/{:.5f}", dx.norm(), _minParameterUpdate);
         break;
       }
-      MLOG(DEBUG) << format("level: {}, iteration: {}, error: {}, #: {},", _level, _iteration, error, constraints[_level].size());
+      MLOG(DEBUG) << format(
+        "level: {}, i: {}, #: {}, err: {}, cov: {}",
+        _level,
+        _iteration,
+        constraints[_level].size(),
+        error,
+        covariance.diagonal().transpose().block(0, 0, 1, 3).norm());
     }
-    MLOG(DEBUG) << format(
-      "Done: {}, level: {}, iteration: {}, error: {}, #: {},", reason, _level, _iteration, error, constraints[_level].size());
+    MLOG(INFO) << format(
+      "Done: {}, level: {}, i: {}, #: {}, err: {}, cov: {},",
+      reason,
+      _level,
+      _iteration,
+      constraints[_level].size(),
+      error,
+      covariance.diagonal().transpose().block(0, 0, 1, 3).norm());
   }
   auto r = std::make_unique<Results>(Results{
     Pose(pose.cast<double>(), covariance.cast<double>()),
     std::vector<Constraint::VecConstShPtr>(constraints.size()),
     scale,
     iterations,
+    _levels,
     normalEquations});
   std::transform(constraints.begin(), constraints.end(), r->constraints.begin(), [](auto c) {
     return Constraint::VecConstShPtr{c.begin(), c.end()};
@@ -159,22 +173,14 @@ AlignmentRgb::Results::UnPtr AlignmentRgb::align(const Feature2D::VecConstShPtr 
   return r;
 }
 
-AlignmentRgb::Constraint::VecShPtr
-AlignmentRgb::setupConstraints(const Frame::VecConstShPtr &framesRef, const Feature2D::VecConstShPtr &features) const {
+AlignmentRgbd::Constraint::VecShPtr
+AlignmentRgbd::setupConstraints(const Frame::VecConstShPtr &framesRef, const Feature2D::VecConstShPtr &features) const {
   TIMED_SCOPE_IF(timer2, format("setupConstraints{}", _level), PERFORMANCE_RGBD_ALIGNMENT);
 
-  // TODO we could also first stack all features and then do one transform..
-
   std::vector<Constraint::ShPtr> constraints(features.size());
-  std::transform(std::execution::par_unseq, features.begin(), features.end(), constraints.begin(), [&](const auto &ft) {
-    const Frame::ConstShPtr frame = ft->frame();
-    const cv::Mat &intensity = frame->intensity(_level);
-    const cv::Mat &dI = frame->dI(_level);
-    const double scale = 1.0 / std::pow(2.0, _level);
-
-    const Vec2d uv = ft->position() * scale;
-    const float i = intensity.at<uint8_t>(uv(1), uv(0));
-    const cv::Vec2f dIuv = dI.at<cv::Vec2f>(uv(1), uv(0));
+  std::transform(exec_policy, features.begin(), features.end(), constraints.begin(), [&](const auto &ft) {
+    const Vec2d uv = ft->position() / std::pow(2.0, _level);
+    auto frame = ft->frame();
 
     auto c = std::make_unique<Constraint>();
     for (c->fId = 0; c->fId < framesRef.size(); c->fId++) {
@@ -182,25 +188,26 @@ AlignmentRgb::setupConstraints(const Frame::VecConstShPtr &framesRef, const Feat
         break;
       }
     }
-    c->uv0 = uv.cast<float>();
-    c->p0 = frame->p3d(uv(1), uv(0), _level).cast<float>();
-    c->i0 = i;
+    c->uv0 = Vec2d(ft->position()).cast<float>();
+    c->p0 = Vec3d(frame->p3d(uv(1), uv(0), _level)).cast<float>();
+    c->iz0 = Vec2f(frame->intensity(uv(1), uv(0), _level) / 255., c->p0.z());
 
-    /*TODO
-    - dont we have to recompute this each iteration? NO we always linearize and differentiate around the identity warp
-    */
-    const Mat<float, 1, 2> JI(dIuv[0], dIuv[1]);
+    // Dont we have to recompute this each iteration? NO we always linearize and differentiate around the identity warp
+    const Mat<float, 1, 2> JI(frame->dIx(uv(1), uv(0), _level), frame->dIy(uv(1), uv(0), _level));
+    const Mat<float, 1, 2> JZ(frame->dZx(uv(1), uv(0), _level), frame->dZy(uv(1), uv(0), _level));
 
-    c->J = JI * jacobian::project_p<float>(c->p0, frame->camera(_level)->fx(), frame->camera(_level)->fy()) *
-           jacobian::transform_se3(SE3f(), c->p0);
+    c->J.row(0) = 1. / 255. * JI * jacobian::project_p<float>(c->p0, frame->camera(_level)->fx(), frame->camera(_level)->fy()) *
+                  jacobian::transform_se3(SE3f(), c->p0);
+    c->JZJw = JZ * jacobian::project_p<float>(c->p0, frame->camera(_level)->fx(), frame->camera(_level)->fy()) *
+              jacobian::transform_se3(SE3f(), c->p0);
     return c;
   });
 
   return constraints;
 }
 
-AlignmentRgb::Constraint::VecShPtr AlignmentRgb::computeResidualsAndJacobian(
-  const AlignmentRgb::Constraint::VecShPtr &constraints, Frame::ConstShPtr f1, const std::vector<SE3f> &motion) const {
+AlignmentRgbd::Constraint::VecShPtr AlignmentRgbd::computeResidualsAndJacobian(
+  const AlignmentRgbd::Constraint::VecShPtr &constraints, Frame::ConstShPtr f1, const std::vector<SE3f> &motion) const {
   TIMED_SCOPE_IF(timer1, format("computeResidualAndJacobian{}", _level), PERFORMANCE_RGBD_ALIGNMENT);
 
   /*Cache some constants for faster loop*/
@@ -218,8 +225,10 @@ AlignmentRgb::Constraint::VecShPtr AlignmentRgb::computeResidualsAndJacobian(
   std::transform(motionInv.begin(), motionInv.end(), Rinv.begin(), [](auto m) { return m.rotationMatrix(); });
   std::vector<Vec3f> tinv(motion.size());
   std::transform(motionInv.begin(), motionInv.end(), tinv.begin(), [](auto m) { return m.translation(); });
+  const Mat3f Kinv = cam->Kinv().cast<float>();
 
   const cv::Mat &I1 = f1->I(_level);
+  const cv::Mat &Z1 = f1->Z(_level);
   const float h = f1->height(_level);
   const float w = f1->width(_level);
   const int bh = std::max<int>(1, (int)(0.01f * h));
@@ -227,52 +236,49 @@ AlignmentRgb::Constraint::VecShPtr AlignmentRgb::computeResidualsAndJacobian(
 
   auto withinImage = [&](const Vec2f &uv) -> bool { return (bw < uv(0) && uv(0) < w - bw && bh < uv(1) && uv(1) < h - bh); };
 
-  std::function<float(const Vec2f &)> sample = [&](const Vec2f &uv) -> float {
-    return interpolate(uv, [&](int v, int u) { return Vec<float, 1>::Constant(I1.at<uint8_t>(v, u)); })(0);
-  };
-  if (f1->hasDepth()) {
-    sample = [&](const Vec2f &uv) -> float {
-      return interpolate(uv, [&](int v, int u) {
-        const float i1 = I1.at<uint8_t>(v, u);
-        const float z1 = f1->depth(_level).at<float>(v, u);
-        return z1 > 0 ? Vec2f(i1, z1) : Vec2f::Constant(NANf);
-      })(0);
-    };
-  }
-
-  std::for_each(std::execution::par_unseq, constraints.begin(), constraints.end(), [&](auto c) {
+  std::for_each(exec_policy, constraints.begin(), constraints.end(), [&](auto c) {
     const size_t &i = c->fId;
     const Vec3f p0t = K * ((R[i] * (c->p0)) + t[i]);
     const Vec2f uv0t = Vec2f(p0t(0), p0t(1)) / p0t(2);
     c->uv1 = uv0t;
-    const float i1w = withinImage(uv0t) ? sample(uv0t) : NANf;
+    const Vec2f iz1w = withinImage(uv0t) ? interpolate(
+                                             uv0t,
+                                             [&](int v, int u) {
+                                               const float i1 = I1.at<uint8_t>(v, u);
+                                               const float z1 = Z1.at<float>(v, u);
+                                               return z1 > 0 ? Vec2f(i1, z1) : Vec2f::Constant(NANf);
+                                             })
+                                         : Vec2f::Constant(NANf);
 
-    c->residual = i1w - c->i0;
+    const Vec3f p1t = ((Rinv[i] * (iz1w(1) * (Kinv * Vec3f(uv0t(0), uv0t(1), 1.0)))) + tinv[i]);
 
-    c->valid = p0t.z() > 0 && std::isfinite(i1w) && std::isfinite(c->residual) && std::isfinite(c->J.norm());
+    c->residual = Vec2f(iz1w(0) / 255.0, p1t.z()) - c->iz0;
+
+    c->J.row(1) = c->JZJw - jacobian::transform_se3(motion[i], c->p0).row(2);
+
+    c->valid = p0t.z() > 0 && std::isfinite(iz1w.norm()) && std::isfinite(c->residual.norm()) && std::isfinite(c->J.norm());
   });
   std::vector<Constraint::ShPtr> constraintsValid;
   std::copy_if(constraints.begin(), constraints.end(), std::back_inserter(constraintsValid), [](auto c) { return c->valid; });
   return constraintsValid;
 }
 
-NormalEquations AlignmentRgb::computeNormalEquations(const std::vector<AlignmentRgb::Constraint::ShPtr> &constraints) const {
+NormalEquations AlignmentRgbd::computeNormalEquations(const std::vector<AlignmentRgbd::Constraint::ShPtr> &constraints) const {
   TIMED_SCOPE_IF(timer2, format("computeNormalEquations{}", _level), PERFORMANCE_RGBD_ALIGNMENT);
+  // TODO why does this give inconsistent results when run in parallel?
   NormalEquations ne = std::transform_reduce(
-    std::execution::par_unseq,
-    std::begin(constraints),
-    std::end(constraints),
-    NormalEquations({Mat6f::Zero(), Vec6f::Zero(), 0.0, 0}),
-    std::plus<NormalEquations>{},
-    [](auto c) {
+    std::execution::seq, constraints.begin(), constraints.end(), NormalEquations{}, std::plus<NormalEquations>{}, [](auto c) {
       return NormalEquations(
-        {c->J * c->weight * c->J.transpose(), c->J * c->weight * c->residual, c->residual * c->weight * c->residual, 1});
+        {c->J.transpose() * c->weight * c->J,
+         c->J.transpose() * c->weight * c->residual,
+         c->residual.transpose() * c->weight * c->residual,
+         1});
     });
 
   return ne;
 }
 
-NormalEquations AlignmentRgb::computeNormalEquations(const Pose &prior, const SE3f &pose) {
+NormalEquations AlignmentRgbd::computeNormalEquations(const Pose &prior, const SE3f &pose) {
   const Mat6f priorInformation = prior.cov().inverse().cast<float>();
   const Vec6f priorError = priorInformation * ((pose * prior.SE3().inverse().cast<float>()).log());
   return {priorInformation, priorError, priorError.norm(), 1};
