+++ src/vslam/src/odometry/NormalEquations.cpp
