+++ src/vslam/src/odometry/NormalEquations.h
