+++ src/vslam/test/bundle_adjustment/AlignmentRgbPoseGraph.cpp
