+++ src/vslam/test/bundle_adjustment/AlignmentRgbPoseGraph.h
