+++ src/vslam/test/bundle_adjustment/overlays.cpp
