+++ src/vslam/test/bundle_adjustment/rgb_pose_graph.cpp
