+++ src/vslam/test/bundle_adjustment/rgb_pose_graph_single_frame.cpp
@@ -22,7 +22,7 @@ using namespace testing;
 #include <thread>
 
 #include "vslam/core.h"
-#include "vslam/direct.h"
+#include "vslam/odometry.h"
 #include "vslam/evaluation.h"
 #include "vslam/features.h"
 #include "vslam/pose_prediction.h"
