+++ src/vslam/test/odometry/rgbd_odometry_entropy.cpp
@@ -7,7 +7,7 @@ using namespace testing;
 #include <thread>
 
 #include "vslam/core.h"
-#include "vslam/direct.h"
+#include "vslam/odometry.h"
 #include "vslam/evaluation.h"
 #include "vslam/features.h"
 #include "vslam/pose_prediction.h"
