+++ src/vslam/test/odometry/rgbd_odometry_simple.cpp
@@ -7,7 +7,7 @@ using namespace testing;
 #include <thread>
 
 #include "vslam/core.h"
-#include "vslam/direct.h"
+#include "vslam/odometry.h"
 #include "vslam/evaluation.h"
 #include "vslam/utils.h"
 
@@ -26,10 +26,10 @@ int main(int argc, char **argv) {
   const int tRmse = 200;
   std::thread thread;
 
-  auto alignment = std::make_shared<AlignmentRgb>(AlignmentRgb::defaultParameters());
+  auto directIcp = std::make_shared<AlignmentRgbd>(AlignmentRgbd::defaultParameters());
+
   log::initialize(outPath, true);
   log::configure(TEST_RESOURCE "/log/");
-  log::config("Features")->show = 1;
   Trajectory::ShPtr traj = std::make_shared<Trajectory>();
   const size_t fEnd = dl->timestamps().size();
   Pose motion;
@@ -47,7 +47,7 @@ int main(int argc, char **argv) {
       cv::waitKey(1);
       {
         TIMED_SCOPE(timer, "computeFrame");
-        motion = alignment->align(dl->cam(), img0, depth0, img, motion.SE3());
+        motion = directIcp->align(dl->cam(), img0, depth0, img, depth, motion.SE3());
       }
       pose = motion * pose;
       traj->append(dl->timestamps()[fId], pose.inverse());
