+++ src/vslam/test/odometry/test_direct_icp.cpp
@@ -7,9 +7,9 @@ using namespace testing;
 #include <thread>
 
 #include "vslam/core.h"
-#include "vslam/direct.h"
 #include "vslam/evaluation.h"
 #include "vslam/features.h"
+#include "vslam/odometry.h"
 #include "vslam/pose_prediction.h"
 #include "vslam/utils.h"
 using namespace vslam;
