+++ src/vslam/test/pose_graph_optimization/loop_closure_detection.cpp
@@ -7,10 +7,10 @@ using namespace testing;
 #include <thread>
 
 #include "vslam/core.h"
-#include "vslam/direct.h"
 #include "vslam/evaluation.h"
 #include "vslam/features.h"
 #include "vslam/loop_closure_detection.h"
+#include "vslam/odometry.h"
 #include "vslam/pose_graph_optimization.h"
 #include "vslam/pose_prediction.h"
 #include "vslam/utils.h"
