+++ src/vslam/tools/Macros.cmake
@@ -73,7 +73,7 @@ macro(pd_add_test name sources)
     list(LENGTH ExtraMacroArgs NumExtraMacroArgs)
     if(NumExtraMacroArgs GREATER 0)
         foreach(ExtraArg ${ExtraMacroArgs})
-            target_compile_definitions(test_${name} PUBLIC ${ExtraArg})
+            target_sources(${name} PRIVATE ${ExtraArg})
         endforeach()
     endif()
 endmacro()
@@ -92,5 +92,12 @@ macro(pd_add_experiment name sources)
     LIBRARY_OUTPUT_DIRECTORY "${VSLAM_TEST_BINARY_DIR}"
     RUNTIME_OUTPUT_DIRECTORY "${VSLAM_TEST_BINARY_DIR}"
     )
+    set(ExtraMacroArgs ${ARGN})
+    list(LENGTH ExtraMacroArgs NumExtraMacroArgs)
+    if(NumExtraMacroArgs GREATER 0)
+        foreach(ExtraArg ${ExtraMacroArgs})
+            target_sources(${name} PRIVATE ${ExtraArg})
+        endforeach()
+    endif()
 
 endmacro()
